+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Using an index store that does not contain complete data for the requested targets now results in an error.
 - The `--index-store-path` option now implies `--skip-build`.
 - Guided setup now omits the `--targets` option from the generated command when 'all' targets are selected for SwiftPM projects.
+- Add option to guided setup to save the configuration to '.periphery.yml'.
 
 ##### Bug Fixes